requirements.txt
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
numpy==2.2.6
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.5
scikit-learn==1.7.1
scipy==1.15.3
sniffio==1.3.1
threadpoolctl==3.6.0
typing_extensions==4.14.1
urllib3==2.5.0
whois==1.20240129.2
